DataWorkstation

Documentation for DataWorkstation.

API Reference

Index

DataWorkstation.Config.ConfigObject

DataWorkstation.IO.FileHandler

DataWorkstation.Config.load_config

DataWorkstation.Config.parse_config

DataWorkstation.Config.update_config

DataWorkstation.IO.custom_logger_meta_formatter

DataWorkstation.IO.get_formatted_logger

DataWorkstation.IO.load_file

DataWorkstation.IO.methods_load_file

DataWorkstation.IO.methods_save_file

DataWorkstation.IO.register_default_bson_file_handler!

DataWorkstation.IO.register_default_csv_file_handler!

DataWorkstation.IO.register_default_jdf_file_handler!

DataWorkstation.IO.register_default_serialization_file_handler!

DataWorkstation.IO.register_file_handler!

DataWorkstation.IO.save_file

DataWorkstation.IO.unregister_file_handler!

Config module

DataWorkstation.Config.ConfigObject — Type

ConfigObject(nt::NamedTuple)
ConfigObject(cfg::ConfigObject)

Abstraction to manage configuration values in a program. Elements in the object being NamedTuple are converted to ConfigObject instances.

Fields

_nt::NamedTuple: stores the keys and values for the configuration

Examples

julia> raw = (;a=1, b=(;c=3, d=4))
(a = 1, b = (c = 3, d = 4))

julia> cfg = ConfigObject(raw)
ConfigObject((a = 1, b = ConfigObject((c = 3, d = 4))))

julia> cfg.b
ConfigObject((c = 3, d = 4))

julia> cfg.b.d
4

julia> length(cfg)
2

julia> keys(cfg), values(cfg)
((:a, :b), (1, ConfigObject((c = 3, d = 4))))

julia> merge(cfg, ConfigObject((;e=5, f=6)))
ConfigObject((a = 1, b = ConfigObject((c = 3, d = 4)), e = 5, f = 6))

julia> cfg == ConfigObject(raw)
true

julia> collect(cfg)
(a = 1, b = (c = 3, d = 4))

DataWorkstation.Config.load_config — Function

load_config(filename, root) -> ConfigObject

Load a file as a ConfigObject instance.

Arguments

filename::AbstractString: the path to load to get the configuration content

root::AbstractString: root for the path to load (optional)

Returns

ConfigObject: having the content loaded from the file

Examples

julia> config_dict = Dict("a" => 1, "b" => Dict("c" => 2, "d" => 3))
Dict{String, Any} with 2 entries:
  "b" => Dict("c"=>2, "d"=>3)
  "a" => 1

julia> using TOML

julia> filename = tempdir() * "/cfg.toml"
"/tmp/cfg.toml"

julia> open(filename, "w") do io
       TOML.print(io, config_dict)
       end;

julia> load_config(filename)
ConfigObject((b = ConfigObject((c = 2, d = 3)), a = 1))

DataWorkstation.Config.parse_config — Function

parse_config(d::Dict, root::AbstractString = "") -> ConfigObject
parse_config(s::AbstractString, root) -> Union{ConfigObject, String}

Parse an input to build a ConfigObject instance. If the input is a Dict, the content will be converted into entries for a ConfigObject. If some entry contains a string of the kind "filename.toml", it is taken as a file to load as a ConfigObject.

Arguments

d::Dict: the dict to convert to a ConfigObject instance

s::AbstractString: when a value is a string with a format like "filename.toml", the file is loaded from that path as a ConfigObject

root::AbstractString="": in case of loading a file, this is the root for the path to load (optional)

Returns

ConfigObject: having the content parsed from the input

Examples

julia> config_dict = Dict("a" => 1, "b" => Dict("c" => 2, "d" => 3))
Dict{String, Any} with 2 entries:
  "b" => Dict("c"=>2, "d"=>3)
  "a" => 1

julia> parse_config(config_dict)
ConfigObject((b = ConfigObject((c = 2, d = 3)), a = 1))

julia> using TOML

julia> filename = tempdir() * "/cfg.toml"
"/tmp/cfg.toml"

julia> open(filename, "w") do io
       TOML.print(io, config_dict)
       end;

julia> parse_config(Dict("cfg" => "\$($(filename))"))
ConfigObject((cfg = ConfigObject((b = ConfigObject((c = 2, d = 3)), a = 1)),))

DataWorkstation.Config.update_config — Function

update_config(cfg::ConfigObject, entries::NamedTuple) -> ConfigObject

Get a new ConfigObject by updating the content with new entries

Arguments

cfg::ConfigObject: configuration instance to updated

entries::NamedTuple: new entries to put into the configuration instance

Returns

ConfigObject: having updated entries

Examples

julia> raw = (;a=1, b=(;c=3, d=4))
(a = 1, b = (c = 3, d = 4))

julia> cfg = ConfigObject(raw)
ConfigObject((a = 1, b = ConfigObject((c = 3, d = 4))))

julia> update_config(cfg, (;e=5, f=6))
ConfigObject((a = 1, b = ConfigObject((c = 3, d = 4)), e = 5, f = 6))

IO module

File operations

DataWorkstation.IO.FileHandler — Type

FileHandler{E,V}
FileHandler(extension::AbstractString, version::Union{AbstractString,Nothing} = nothing)

Abstraction to handle file operations. It is internally managed by load_file() and save_file() methods, so it should not be necessary to instantiate it directly.

Examples

julia> fh = FileHandler("csv", "dataframe")
FileHandler{:csv, :dataframe}()

julia> FileHandler{:csv, :dataframe} == typeof(fh)
true

julia> FileHandler("csv")
FileHandler{:csv, :nothing}()

DataWorkstation.IO.load_file — Function

load_file(
    path::AbstractString,
    args...;
    version::Union{AbstractString,Nothing} = nothing,
    kwargs...,
) -> Any

Function to load the content of file in path, using a method already registered with a call of register_file_handler!() with version and the file extension given in path.

Arguments

path::AbstractString: path to the file to load, which should have an extension compatible with a load_file() method already registered

args...: Arguments passed to the registered load_file() method

Keywords

version::Union{AbstractString,Nothing}: tag to identify the particular method to use

kwargs...: Keyword arguments passed to the registered load_file() method

Returns

Any: object loaded from the file, and the type will depend on the method implemented

Throws

ErrorException: In case there is no load_file() method compatible with version and the extension given in path, or if some error occurs during the load operation.

Examples

julia> using CSV, DataFrames

julia> register_file_handler!(
            "csv",
            load_file_function = path -> CSV.read(path, DataFrame),
            save_file_function = (path, obj) -> CSV.write(path, obj),
        );

julia> df = DataFrame(Dict("a" => 1:5, "b" => 6:10))
5×2 DataFrame
 Row │ a      b
     │ Int64  Int64
─────┼──────────────
   1 │     1      6
   2 │     2      7
   3 │     3      8
   4 │     4      9
   5 │     5     10

julia> save_file("dataset.csv", df);

julia> isfile("dataset.csv")
true

julia> run(`head -2 dataset.csv`)
a,b
1,6
Process(`head -2 dataset.csv`, ProcessExited(0))

julia> load_file("dataset.csv")
5×2 DataFrame
 Row │ a      b
     │ Int64  Int64
─────┼──────────────
   1 │     1      6
   2 │     2      7
   3 │     3      8
   4 │     4      9
   5 │     5     10

julia> register_file_handler!(
            "csv",
            version = "pipe_delim",
            load_file_function = path -> CSV.read(path, DataFrame, delim="|"),
            save_file_function = (path, obj) -> CSV.write(path, obj, delim="|"),
        );

julia> save_file("dataset_pipe.csv", df, version="pipe_delim");

julia> run(`head -2 dataset_pipe.csv`)
a|b
1|6
Process(`head -2 dataset_pipe.csv`, ProcessExited(0))

julia> load_file("dataset_pipe.csv", version="pipe_delim")
5×2 DataFrame
 Row │ a      b
     │ Int64  Int64
─────┼──────────────
   1 │     1      6
   2 │     2      7
   3 │     3      8
   4 │     4      9
   5 │     5     10

julia> register_file_handler!(
            "csv",
            version = "verbose",
            load_file_function = (path; verbose=false) -> begin
                verbose ? println("Loading dataset from $path...") : nothing
                CSV.read(path, DataFrame)
            end,
            save_file_function = (path, obj; verbose=false) -> begin
                verbose ? println("Saving dataset to $path...") : nothing
                CSV.write(path, obj)
            end
        );

julia> save_file("dataset.csv", df, version="verbose", verbose=true);
Saving dataset to dataset.csv...

julia> load_file("dataset.csv", version="verbose", verbose=true)
Loading dataset from dataset.csv...
5×2 DataFrame
 Row │ a      b
     │ Int64  Int64
─────┼──────────────
   1 │     1      6
   2 │     2      7
   3 │     3      8
   4 │     4      9
   5 │     5     10

julia> length(methods_load_file("csv", nothing))
1

julia> unregister_file_handler!("csv", version=nothing);

julia> length(methods_load_file("csv", nothing))
0

DataWorkstation.IO.save_file — Function

save_file(
    path::AbstractString,
    obj::Any,
    args...;
    version::Union{AbstractString,Nothing} = nothing,
    kwargs...,
) -> Nothing

Function to save the content of obj as a file in path, using a method already registered with a call of register_file_handler!() with version and the file extension given in path.

For examples of usage, look at the documentation of load_file().

Arguments

path::AbstractString: path to the file to save, which should have an extension compatible with a save_file() method already registered

obj::Any: object to be saved as a file in the given path

args...: Arguments passed to the registered save_file() method

Keywords

version::Union{AbstractString,Nothing}: tag to identify the particular method to use

kwargs...: Keyword arguments passed to the registered save_file() method

Throws

ErrorException: In case there is no save_file() method compatible with version and the extension given in path, or if some error occurs during the save operation.

DataWorkstation.IO.methods_load_file — Function

methods_load_file(
    extension::AbstractString,
    version::Union{AbstractString,Nothing} = nothing
) -> Vector{Method}

Get the available methods for load_file() that are compatible with the arguments extension and version. This is useful to check if there are methods already registered with register_file_handler!() that support the arguments.

Arguments

extension::AbstractString: extension of a file path to support in the registered method

version::Union{AbstractString,Nothing}: version of the registered methods

Returns

Vector{Method}: available methods that are compatible with the given arguments

Examples

julia> using CSV, DataFrames

julia> length(methods_load_file("csv", nothing))
0

julia> register_file_handler!(
            "csv",
            load_file_function = path -> CSV.read(path, DataFrame),
            save_file_function = (path, obj) -> CSV.write(path, obj),
        );

julia> length(methods_load_file("csv", nothing))
1

julia> unregister_file_handler!("csv", version=nothing);

julia> length(methods_load_file("csv", nothing))
0

DataWorkstation.IO.methods_save_file — Function

methods_save_file(
    extension::AbstractString,
    version::Union{AbstractString,Nothing} = nothing
) -> Vector{Method}

Get the available methods for save_file() that are compatible with the arguments extension and version. This is useful to check if there are methods already registered with register_file_handler!() that support the arguments.

Arguments

extension::AbstractString: extension of a file path to support in the registered method

version::Union{AbstractString,Nothing}: version of the registered methods

Returns

Vector{Method}: available methods that are compatible with the given arguments

Examples

julia> using CSV, DataFrames

julia> length(methods_save_file("csv", nothing))
0

julia> register_file_handler!(
            "csv",
            load_file_function = path -> CSV.read(path, DataFrame),
            save_file_function = (path, obj) -> CSV.write(path, obj),
        );

julia> length(methods_save_file("csv", nothing))
1

julia> unregister_file_handler!("csv", version=nothing);

julia> length(methods_save_file("csv", nothing))
0

DataWorkstation.IO.register_file_handler! — Function

register_file_handler!(
    extension::AbstractString;
    version::Union{AbstractString,Nothing} = nothing,
    load_file_function::Function,
    save_file_function::Function,
) -> Nothing

Function to register methods for load_file() and save_file() that are compatible with file paths having an extension value and some version.

This way allows to use a single function to save any object with a proper method just based on the extension of the given filename, which is useful to avoid hard-coded file operations in your program.

You can handle multiple methods for a given extension, using the argument version to specify which method to use for a given file.

For examples of usage, look at the documentation of load_file().

Arguments

extension::AbstractString: extension to extract from a file path to recognize that the methods registered should be used

Keywords

version::Union{AbstractString,Nothing}: tag to identify the type of methods to register

load_file_function::Function: function to use in the call of load_file()

save_file_function::Function: function to use in the call of save_file()

Throws

ErrorException: In case there are already load_file() and/or save_file() methods compatible with version and the given extension

DataWorkstation.IO.unregister_file_handler! — Function

unregister_file_handler!(
    extension::String;
    version::Union{AbstractString,Nothing} = nothing
) -> Nothing

Function to unregister available methods for load_file() and save_file() that were already registered with a call of register_file_handler!() having an extension value and some version.

For examples of usage, look at the documentation of load_file().

Arguments

extension::AbstractString: extension to extract from a file path to recognize the methods to unregister

Keywords

version::Union{AbstractString,Nothing}: tag for the type of methods to unregister

Throws

ErrorException: In case there are no load_file() or save_file() methods compatible with version and the given extension to unregister

Util file handlers

DataWorkstation.IO.register_default_bson_file_handler! — Function

register_default_bson_file_handler!(; extension = "bson", version = nothing) -> Nothing

Function to register methods for load_file() and save_file() with a default implementation based on the BSON library, to be compatible with file paths having an extension value and some version. This internall calls register_file_handler!().

Useful to save objects like machine learning models or just Julia objects to be retrieved in a Julia session, using a BSON format.

Arguments

extension::AbstractString: extension of a file path to support in the registered method

version::Union{AbstractString,Nothing}: version of the registered methods

Examples

julia> extension = "bson";

julia> version = nothing;

julia> obj = (;a=1, b=2);

julia> filepath = joinpath(tempdir(), "file.$extension");

julia> register_default_bson_file_handler!(
            extension = extension;
            version = version,
        );

julia> save_file(filepath, obj, version=version);

julia> loaded_obj = load_file(filepath, version=version);

julia> loaded_obj == obj
true

DataWorkstation.IO.register_default_csv_file_handler! — Function

register_default_csv_file_handler!(; extension = "csv", version = nothing) -> Nothing

Function to register methods for load_file() and save_file() with a default implementation based on the CSV library, to be compatible with file paths having an extension value and some version. This internall calls register_file_handler!().

Useful to save tabular data in plain-text format. Make sure the object is in a compatible format before saving it (e.g. Array, DataFrame, Dict).

Arguments

extension::AbstractString: extension of a file path to support in the registered method

version::Union{AbstractString,Nothing}: version of the registered methods

Examples

julia> using DataFrames

julia> extension = "csv";

julia> version = "v1";

julia> df = DataFrame(Dict("a" => 1:5, "b" => 6:10))
5×2 DataFrame
 Row │ a      b
     │ Int64  Int64
─────┼──────────────
   1 │     1      6
   2 │     2      7
   3 │     3      8
   4 │     4      9
   5 │     5     10

julia> filepath = joinpath(tempdir(), "file.$extension");

julia> register_default_csv_file_handler!(
            extension = extension;
            version = version,
        );

julia> save_file(filepath, df, version=version);

julia> loaded_df = load_file(filepath, version=version);

julia> loaded_df == df
true

DataWorkstation.IO.register_default_jdf_file_handler! — Function

register_default_jdf_file_handler!(; extension = "jdf", version = nothing) -> Nothing

Function to register methods for load_file() and save_file() with a default implementation based on the JDF library, to be compatible with file paths having an extension value and some version. This internall calls register_file_handler!().

Useful to save tabular data in binary format. Make sure the object is in DataFrame format.

Arguments

extension::AbstractString: extension of a file path to support in the registered method

version::Union{AbstractString,Nothing}: version of the registered methods

Examples

julia> using DataFrames

julia> extension = "jdf";

julia> version = nothing;

julia> df = DataFrame(Dict("a" => 1:5, "b" => 6:10))
5×2 DataFrame
 Row │ a      b
     │ Int64  Int64
─────┼──────────────
   1 │     1      6
   2 │     2      7
   3 │     3      8
   4 │     4      9
   5 │     5     10

julia> filepath = joinpath(tempdir(), "file.$extension");

julia> register_default_jdf_file_handler!(
            extension = extension;
            version = version,
        );

julia> save_file(filepath, df, version=version);

julia> loaded_df = load_file(filepath, version=version);

julia> loaded_df == df
true

DataWorkstation.IO.register_default_serialization_file_handler! — Function

register_default_serialization_file_handler!(
    ; extension = "slz", version = nothing) -> Nothing

Function to register methods for load_file() and save_file() with a default implementation based on the Serialization library, to be compatible with file paths having an extension value and some version. This internall calls register_file_handler!().

Useful to save objects like machine learning models or just Julia objects to be retrieved in a Julia session, by using a native Julia implementation.

Arguments

extension::AbstractString: extension of a file path to support in the registered method

version::Union{AbstractString,Nothing}: version of the registered methods

Examples

julia> extension = "slz";

julia> version = nothing;

julia> obj = (;a=1, b=2);

julia> filepath = joinpath(tempdir(), "file.$extension");

julia> register_default_serialization_file_handler!(
            extension = extension;
            version = version,
        );

julia> save_file(filepath, obj, version=version);

julia> loaded_obj = load_file(filepath, version=version);

julia> loaded_obj == obj
true

Logging utils

DataWorkstation.IO.custom_logger_meta_formatter — Function

custom_logger_meta_formatter(
    labels::Union{Vector{T}, Nothing} = nothing,
    log_level::Logging.LogLevel = Logging.Debug;
    sep::AbstractString = " | ",
    date_format::Union{AbstractString, Nothing} = "yyyy-mm-dd HH:MM:SS",
) where {T <: Union{Any, String}} -> Tuple{Symbol, String, String}

Function to allow a custom format for the logging messages to generate with a Logging.ConsoleLogger instance. The format is the following one: [ (date) (sep) (labels separated by 'sep') (log level): (message) For example: [ 2022-01-01 00:00:00 | tag=test | INFO: Happy new year!

Arguments

labels::Union{Vector{T}, Nothing}: labels to be used in the prefix of the logging messages. If nothing or an empty Array, no labels will be added to the message.

log_level::Logging.LogLevel: Level for the log message to print.

Keywords

sep::AbstractString: Separator for the labels to print in the message prefix

date_format::Union{AbstractString, Nothing}: Format for the timestamp to print in the message prefix. If nothing, no timestamp will be printed.

Returns

Tuple{Symbol, String, String}: A tuple of log level, prefix of message, and suffix of message to be used by a Logger instance when printing a logging message.

Examples

julia> using Logging

julia> custom_logger_meta_formatter(date_format=nothing)
(:blue, "DEBUG:", "")

julia> custom_logger_meta_formatter(["level", "sublevel"],
       Logging.Info, date_format=nothing, sep=" - ")
(:cyan, "level - sublevel - INFO:", "")

DataWorkstation.IO.get_formatted_logger — Function

get_formatted_logger(
    min_level::Symbol=:Debug,
    args...;
    stream::Base.IO=stderr,
    sep::String=" | ",
    date_format::AbstractString = "yyyy-mm-dd HH:MM:SS",
    kwargs...) -> Logging.ConsoleLogger

Get an instance of Logging.ConsoleLogger, having a format for the messages based on the given arguments.

This is useful for the package to log the operations on a workflow with details about in which level they occur, which is important during debugging.

Arguments

min_level::Symbol: Minimum level of log to print, represented as as symbol (e.g. :Debug, :Info, :Error, :Warn).

args...: Labels to be printed as in the message prefix.

Keywords

stream::Base.IO: Stream for the messages to print. Default to stderr.

sep::AbstractString: Separator for the labels to print in the message prefix.

date_format::Union{AbstractString, Nothing}: Format for the timestamp. to print in the message prefix. If nothing, no timestamp will be printed.

kwargs...: Labels to be printed as key=value in the message prefix.

Returns

Logging.ConsoleLogger: Logger formatted with custom_logger_meta_formatter().

Throws

ErrorException: In case the log level is not supported, or if some error occurs during the logger obtention.

julia> using Logging

julia> Logging.with_logger(get_formatted_logger(scope="repl", tag="test")) do
            @info "Hello world!"
        end
[ 2022-05-18 17:27:01 | scope=repl | tag=test | INFO: Hello world!

julia> Logging.with_logger(get_formatted_logger(:Debug, "tag1", level="1")) do
            @debug "Logging in this level"
            Logging.with_logger(get_formatted_logger(:Debug, "tag2", level="2")) do
                @debug "Logging in another level"
            end
        end
[ 2022-05-18 17:27:16 | tag1 | level=1 | DEBUG: Logging in this level
[ 2022-05-18 17:27:16 | tag2 | level=2 | DEBUG: Logging in another level