DataWorkstation

Documentation for DataWorkstation.

API Reference

Index

DataWorkstation.Config.ConfigObject

DataWorkstation.Config.load_config

DataWorkstation.Config.parse_config

DataWorkstation.Config.update_config

Config module

DataWorkstation.Config.ConfigObject — Type

ConfigObject(nt::NamedTuple)
ConfigObject(cfg::ConfigObject)

Abstraction to manage configuration values in a program. Elements in the object being NamedTuple are converted to ConfigObject instances.

Fields

_nt::NamedTuple: stores the keys and values for the configuration

Examples

julia> raw = (;a=1, b=(;c=3, d=4))
(a = 1, b = (c = 3, d = 4))

julia> cfg = ConfigObject(raw)
ConfigObject((a = 1, b = ConfigObject((c = 3, d = 4))))

julia> cfg.b
ConfigObject((c = 3, d = 4))

julia> cfg.b.d
4

julia> length(cfg)
2

julia> keys(cfg), values(cfg)
((:a, :b), (1, ConfigObject((c = 3, d = 4))))

julia> merge(cfg, (;e=5, f=6))
ConfigObject((a = 1, b = ConfigObject((c = 3, d = 4)), e = 5, f = 6))

julia> cfg == ConfigObject(raw)
true

julia> collect(cfg)
(a = 1, b = (c = 3, d = 4))

DataWorkstation.Config.load_config — Function

load_config(filename, root) -> ConfigObject

Load a file as a ConfigObject instance.

Arguments

filename::AbstractString: the path to load to get the configuration content

root::AbstractString: root for the path to load (optional)

Returns

ConfigObject: having the content loaded from the file

Examples

julia> config_dict = Dict("a" => 1, "b" => Dict("c" => 2, "d" => 3))
Dict{String, Any} with 2 entries:
  "b" => Dict("c"=>2, "d"=>3)
  "a" => 1

julia> using TOML

julia> filename = tempdir() * "/cfg.toml"
"/tmp/cfg.toml"

julia> open(filename, "w") do io
       TOML.print(io, config_dict)
       end;

julia> load_config(filename)
ConfigObject((b = ConfigObject((c = 2, d = 3)), a = 1))

DataWorkstation.Config.parse_config — Function

parse_config(d::Dict, root::AbstractString = "") -> ConfigObject
parse_config(s::AbstractString, root) -> Union{ConfigObject, String}

Parse an input to build a ConfigObject instance. If the input is a Dict, the content will be converted into entries for a ConfigObject. If some entry contains a string of the kind "filename.toml", it is taken as a file to load as a ConfigObject.

Arguments

d::Dict: the dict to convert to a ConfigObject instance

s::AbstractString: when a value is a string with a format like "filename.toml", the file is loaded from that path as a ConfigObject

root::AbstractString="": in case of loading a file, this is the root for the path to load (optional)

Returns

ConfigObject: having the content parsed from the input

Examples

julia> config_dict = Dict("a" => 1, "b" => Dict("c" => 2, "d" => 3))
Dict{String, Any} with 2 entries:
  "b" => Dict("c"=>2, "d"=>3)
  "a" => 1

julia> parse_config(config_dict)
ConfigObject((b = ConfigObject((c = 2, d = 3)), a = 1))

julia> using TOML

julia> filename = tempdir() * "/cfg.toml"
"/tmp/cfg.toml"

julia> open(filename, "w") do io
       TOML.print(io, config_dict)
       end;

julia> parse_config(Dict("cfg" => "\$($(filename))"))
ConfigObject((cfg = ConfigObject((b = ConfigObject((c = 2, d = 3)), a = 1)),))

DataWorkstation.Config.update_config — Function

update_config(cfg::ConfigObject, entries::NamedTuple) -> ConfigObject

Get a new ConfigObject by updating the content with new entries

Arguments

cfg::ConfigObject: configuration instance to updated

entries::NamedTuple: new entries to put into the configuration instance

Returns

ConfigObject: having updated entries

Examples

julia> raw = (;a=1, b=(;c=3, d=4))
(a = 1, b = (c = 3, d = 4))

julia> cfg = ConfigObject(raw)
ConfigObject((a = 1, b = ConfigObject((c = 3, d = 4))))

julia> update_config(cfg, (;e=5, f=6))
ConfigObject((a = 1, b = ConfigObject((c = 3, d = 4)), e = 5, f = 6))