DataWorkstation

Documentation for DataWorkstation.

API Reference

Index

DataWorkstation.Config.ConfigObject

DataWorkstation.IO.ColorsConfig

DataWorkstation.IO.FileHandler

DataWorkstation.Workflows.LoggingConfig

DataWorkstation.Workflows.WorkflowSpec

DataWorkstation.Config.load_config

DataWorkstation.Config.parse_config

DataWorkstation.Config.update_config

DataWorkstation.IO.custom_logger_meta_formatter

DataWorkstation.IO.get_formatted_logger

DataWorkstation.IO.load_file

DataWorkstation.IO.methods_load_file

DataWorkstation.IO.methods_save_file

DataWorkstation.IO.register_default_bson_file_handler

DataWorkstation.IO.register_default_csv_file_handler

DataWorkstation.IO.register_default_jdf_file_handler

DataWorkstation.IO.register_default_serialization_file_handler

DataWorkstation.IO.register_file_handler

DataWorkstation.IO.save_file

DataWorkstation.IO.unregister_file_handler

DataWorkstation.Workflows.is_valid_step_function

DataWorkstation.Workflows.register_step_function

DataWorkstation.Workflows.run_step_function

DataWorkstation.Workflows.run_workflow

Config module

DataWorkstation.Config.ConfigObject — Type

ConfigObject()
ConfigObject(nt::NamedTuple)
ConfigObject(cfg::ConfigObject)

Abstraction to manage configuration values in a program. Elements in the object that are NamedTuples will be converted to ConfigObject instances. Notice that the manipulation of a ConfigObject is the same as with a NamedTuple. If ConfigObject is called without arguments, it will be created with an empty NamedTuple.

Fields

_nt::NamedTuple: stores the keys and values for the configuration

Throws

ErrorException: In case there is a '_nt' entry in the passed NamedTuple (which is an illegal name for the entries), or if some error occurs during the creation.

Examples

julia> raw = (;a=1, b=(;c=3, d=4))
(a = 1, b = (c = 3, d = 4))

julia> cfg = ConfigObject(raw)
ConfigObject((a = 1, b = ConfigObject((c = 3, d = 4))))

julia> cfg.b
ConfigObject((c = 3, d = 4))

julia> cfg.b.d
4

julia> length(cfg)
2

julia> keys(cfg), values(cfg)
((:a, :b), (1, ConfigObject((c = 3, d = 4))))

julia> merge(cfg, ConfigObject((;e=5, f=6)))
ConfigObject((a = 1, b = ConfigObject((c = 3, d = 4)), e = 5, f = 6))

julia> cfg == ConfigObject(raw)
true

julia> collect(cfg)
(a = 1, b = (c = 3, d = 4))

DataWorkstation.Config.load_config — Function

load_config(filename, root) -> ConfigObject

Load a file as a ConfigObject instance.

Arguments

filename::AbstractString: the path to load to get the configuration content

root::Union{AbstractString,Nothing}: root for the path to load (optional)

Returns

ConfigObject: having the content loaded from the file

Examples

julia> config_dict = Dict("a" => 1, "b" => Dict("c" => 2, "d" => 3))
Dict{String, Any} with 2 entries:
  "b" => Dict("c"=>2, "d"=>3)
  "a" => 1

julia> using TOML

julia> filename = tempdir() * "/cfg.toml"
"/tmp/cfg.toml"

julia> open(filename, "w") do io
       TOML.print(io, config_dict)
       end;

julia> load_config(filename)
ConfigObject((b = ConfigObject((c = 2, d = 3)), a = 1))

DataWorkstation.Config.parse_config — Function

parse_config(d::Dict, root::Union{AbstractString,Nothing}) -> ConfigObject
parse_config(s::AbstractString, root) -> Union{ConfigObject, String}

Parse an input to build a ConfigObject instance. If the input is a Dict, the content will be converted into entries for a ConfigObject. If some entry contains a string of the kind "filename.toml", it is taken as a file to load as a ConfigObject.

Arguments

d::Dict: the dict to convert to a ConfigObject instance

s::AbstractString: when a value is a string with a format like "filename.toml", the file is loaded from that path as a ConfigObject

root::Union{AbstractString,Nothing}: in case of loading a file, this is the root for the path to load (optional)

Returns

ConfigObject: having the content parsed from the input

Examples

julia> config_dict = Dict("a" => 1, "b" => Dict("c" => 2, "d" => 3))
Dict{String, Any} with 2 entries:
  "b" => Dict("c"=>2, "d"=>3)
  "a" => 1

julia> parse_config(config_dict)
ConfigObject((b = ConfigObject((c = 2, d = 3)), a = 1))

julia> using TOML

julia> filename = tempdir() * "/cfg.toml";

julia> open(filename, "w") do io
       TOML.print(io, config_dict)
       end;

julia> parse_config(Dict("cfg" => "\$($(filename))"))
ConfigObject((cfg = ConfigObject((b = ConfigObject((c = 2, d = 3)), a = 1)),))

DataWorkstation.Config.update_config — Function

update_config(cfg::ConfigObject, entries::NamedTuple) -> ConfigObject

Get a new ConfigObject by updating the content with new entries.

!!! note This function does not modify the argument cfg.

Arguments

cfg::ConfigObject: configuration instance to updated

entries::NamedTuple: new entries to put into the configuration instance

Returns

ConfigObject: having updated entries

Examples

julia> raw = (;a=1, b=(;c=3, d=4))
(a = 1, b = (c = 3, d = 4))

julia> cfg = ConfigObject(raw)
ConfigObject((a = 1, b = ConfigObject((c = 3, d = 4))))

julia> update_config(cfg, (;e=5, f=6))
ConfigObject((a = 1, b = ConfigObject((c = 3, d = 4)), e = 5, f = 6))

IO module

File operations

DataWorkstation.IO.FileHandler — Type

FileHandler{E,V}
FileHandler(extension::AbstractString, version::Union{AbstractString,Nothing} = nothing)

Abstraction to handle file operations. It is internally managed by load_file() and save_file() methods, so it should not be necessary to instantiate it directly.

Examples

julia> fh = FileHandler("csv", "dataframe")
FileHandler{:csv, :dataframe}()

julia> FileHandler{:csv, :dataframe} == typeof(fh)
true

julia> FileHandler("csv")
FileHandler{:csv, :nothing}()

DataWorkstation.IO.load_file — Function

load_file(
    path::AbstractString,
    args...;
    version::Union{AbstractString,Nothing} = nothing,
    kwargs...,
) -> Any

Function to load the content of file in path, using a method already registered with a call of register_file_handler() with version and the file extension given in path.

Arguments

path::AbstractString: path to the file to load, which should have an extension compatible with a load_file() method already registered

args...: Arguments passed to the registered load_file() method

Keywords

version::Union{AbstractString,Nothing}: tag to identify the particular method to use

kwargs...: Keyword arguments passed to the registered load_file() method

Returns

Any: object loaded from the file, and the type will depend on the method implemented

Throws

ErrorException: In case there is no load_file() method compatible with version and the extension given in path, or if some error occurs during the load operation.

Examples

julia> using CSV, DataFrames

julia> register_file_handler(
            "csv",
            load_file_function = path -> CSV.read(path, DataFrame),
            save_file_function = (path, obj) -> CSV.write(path, obj),
        );

julia> df = DataFrame(Dict("a" => 1:5, "b" => 6:10))
5×2 DataFrame
 Row │ a      b
     │ Int64  Int64
─────┼──────────────
   1 │     1      6
   2 │     2      7
   3 │     3      8
   4 │     4      9
   5 │     5     10

julia> save_file("dataset.csv", df);

julia> isfile("dataset.csv")
true

julia> run(`head -2 dataset.csv`)
a,b
1,6
Process(`head -2 dataset.csv`, ProcessExited(0))

julia> load_file("dataset.csv")
5×2 DataFrame
 Row │ a      b
     │ Int64  Int64
─────┼──────────────
   1 │     1      6
   2 │     2      7
   3 │     3      8
   4 │     4      9
   5 │     5     10

julia> register_file_handler(
            "csv",
            version = "pipe_delim",
            load_file_function = path -> CSV.read(path, DataFrame, delim="|"),
            save_file_function = (path, obj) -> CSV.write(path, obj, delim="|"),
        );

julia> save_file("dataset_pipe.csv", df, version="pipe_delim");

julia> run(`head -2 dataset_pipe.csv`)
a|b
1|6
Process(`head -2 dataset_pipe.csv`, ProcessExited(0))

julia> load_file("dataset_pipe.csv", version="pipe_delim")
5×2 DataFrame
 Row │ a      b
     │ Int64  Int64
─────┼──────────────
   1 │     1      6
   2 │     2      7
   3 │     3      8
   4 │     4      9
   5 │     5     10

julia> register_file_handler(
            "csv",
            version = "verbose",
            load_file_function = (path; verbose=false) -> begin
                verbose ? println("Loading dataset from $path...") : nothing
                CSV.read(path, DataFrame)
            end,
            save_file_function = (path, obj; verbose=false) -> begin
                verbose ? println("Saving dataset to $path...") : nothing
                CSV.write(path, obj)
            end
        );

julia> save_file("dataset.csv", df, version="verbose", verbose=true);
Saving dataset to dataset.csv...

julia> load_file("dataset.csv", version="verbose", verbose=true)
Loading dataset from dataset.csv...
5×2 DataFrame
 Row │ a      b
     │ Int64  Int64
─────┼──────────────
   1 │     1      6
   2 │     2      7
   3 │     3      8
   4 │     4      9
   5 │     5     10

julia> length(methods_load_file("csv", nothing))
1

julia> unregister_file_handler("csv", version=nothing);

julia> length(methods_load_file("csv", nothing))
0

DataWorkstation.IO.save_file — Function

save_file(
    path::AbstractString,
    obj::Any,
    args...;
    version::Union{AbstractString,Nothing} = nothing,
    kwargs...,
) -> Nothing

Function to save the content of obj as a file in path, using a method already registered with a call of register_file_handler() with version and the file extension given in path.

For examples of usage, look at the documentation of load_file().

Arguments

path::AbstractString: path to the file to save, which should have an extension compatible with a save_file() method already registered

obj::Any: object to be saved as a file in the given path

args...: Arguments passed to the registered save_file() method

Keywords

version::Union{AbstractString,Nothing}: tag to identify the particular method to use

kwargs...: Keyword arguments passed to the registered save_file() method

Throws

ErrorException: In case there is no save_file() method compatible with version and the extension given in path, or if some error occurs during the save operation.

DataWorkstation.IO.methods_load_file — Function

methods_load_file(
    extension::AbstractString,
    version::Union{AbstractString,Nothing} = nothing
) -> Vector{Method}

Get the available methods for load_file() that are compatible with the arguments extension and version. This is useful to check if there are methods already registered with register_file_handler() that support the arguments.

Arguments

extension::AbstractString: extension of a file path to support in the registered method

version::Union{AbstractString,Nothing}: version of the registered methods

Returns

Vector{Method}: available methods that are compatible with the given arguments

Examples

julia> using CSV, DataFrames

julia> length(methods_load_file("csv", nothing))
0

julia> register_file_handler(
            "csv",
            load_file_function = path -> CSV.read(path, DataFrame),
            save_file_function = (path, obj) -> CSV.write(path, obj),
        );

julia> length(methods_load_file("csv", nothing))
1

julia> unregister_file_handler("csv", version=nothing);

julia> length(methods_load_file("csv", nothing))
0

DataWorkstation.IO.methods_save_file — Function

methods_save_file(
    extension::AbstractString,
    version::Union{AbstractString,Nothing} = nothing
) -> Vector{Method}

Get the available methods for save_file() that are compatible with the arguments extension and version. This is useful to check if there are methods already registered with register_file_handler() that support the arguments.

Arguments

extension::AbstractString: extension of a file path to support in the registered method

version::Union{AbstractString,Nothing}: version of the registered methods

Returns

Vector{Method}: available methods that are compatible with the given arguments

Examples

julia> using CSV, DataFrames

julia> length(methods_save_file("csv", nothing))
0

julia> register_file_handler(
            "csv",
            load_file_function = path -> CSV.read(path, DataFrame),
            save_file_function = (path, obj) -> CSV.write(path, obj),
        );

julia> length(methods_save_file("csv", nothing))
1

julia> unregister_file_handler("csv", version=nothing);

julia> length(methods_save_file("csv", nothing))
0

DataWorkstation.IO.register_file_handler — Function

register_file_handler(
    extension::AbstractString;
    version::Union{AbstractString,Nothing} = nothing,
    load_file_function::Function,
    save_file_function::Function,
) -> Nothing

Function to register methods for load_file() and save_file() that are compatible with file paths having an extension value and some version.

This way allows to use a single function to save any object with a proper method just based on the extension of the given filename, which is useful to avoid hard-coded file operations in your program.

You can handle multiple methods for a given extension, using the argument version to specify which method to use for a given file.

For examples of usage, look at the documentation of load_file().

Arguments

extension::AbstractString: extension to extract from a file path to recognize that the methods registered should be used

Keywords

version::Union{AbstractString,Nothing}: tag to identify the type of methods to register

load_file_function::Function: function to use in the call of load_file()

save_file_function::Function: function to use in the call of save_file()

Throws

ErrorException: In case there are already load_file() and/or save_file() methods compatible with version and the given extension

DataWorkstation.IO.unregister_file_handler — Function

unregister_file_handler(
    extension::String;
    version::Union{AbstractString,Nothing} = nothing
) -> Nothing

Function to unregister available methods for load_file() and save_file() that were already registered with a call of register_file_handler() having an extension value and some version.

For examples of usage, look at the documentation of load_file().

Arguments

extension::AbstractString: extension to extract from a file path to recognize the methods to unregister

Keywords

version::Union{AbstractString,Nothing}: tag for the type of methods to unregister

Throws

ErrorException: In case there are no load_file() or save_file() methods compatible with version and the given extension to unregister

Util file handlers

DataWorkstation.IO.register_default_bson_file_handler — Function

register_default_bson_file_handler(; extension = "bson", version = nothing) -> Nothing

Function to register methods for load_file() and save_file() with a default implementation based on the BSON library, to be compatible with file paths having an extension value and some version. This internall calls register_file_handler().

Useful to save objects like machine learning models or just Julia objects to be retrieved in a Julia session, using a BSON format.

Arguments

extension::AbstractString: extension of a file path to support in the registered method

version::Union{AbstractString,Nothing}: version of the registered methods

Examples

julia> extension = "bson";

julia> version = nothing;

julia> obj = (;a=1, b=2);

julia> filepath = joinpath(tempdir(), "file.$extension");

julia> register_default_bson_file_handler(
            extension = extension;
            version = version,
        );

julia> save_file(filepath, obj, version=version);

julia> loaded_obj = load_file(filepath, version=version);

julia> loaded_obj == obj
true

DataWorkstation.IO.register_default_csv_file_handler — Function

register_default_csv_file_handler(; extension = "csv", version = nothing) -> Nothing

Function to register methods for load_file() and save_file() with a default implementation based on the CSV library, to be compatible with file paths having an extension value and some version. This internall calls register_file_handler().

Useful to save tabular data in plain-text format. Make sure the object is in a compatible format before saving it (e.g. Array, DataFrame, Dict).

Arguments

extension::AbstractString: extension of a file path to support in the registered method

version::Union{AbstractString,Nothing}: version of the registered methods

Examples

julia> using DataFrames

julia> extension = "csv";

julia> version = "v1";

julia> df = DataFrame(Dict("a" => 1:5, "b" => 6:10))
5×2 DataFrame
 Row │ a      b
     │ Int64  Int64
─────┼──────────────
   1 │     1      6
   2 │     2      7
   3 │     3      8
   4 │     4      9
   5 │     5     10

julia> filepath = joinpath(tempdir(), "file.$extension");

julia> register_default_csv_file_handler(
            extension = extension;
            version = version,
        );

julia> save_file(filepath, df, version=version);

julia> loaded_df = load_file(filepath, version=version);

julia> loaded_df == df
true

DataWorkstation.IO.register_default_jdf_file_handler — Function

register_default_jdf_file_handler(; extension = "jdf", version = nothing) -> Nothing

Function to register methods for load_file() and save_file() with a default implementation based on the JDF library, to be compatible with file paths having an extension value and some version. This internall calls register_file_handler().

Useful to save tabular data in binary format. Make sure the object is in DataFrame format.

Arguments

extension::AbstractString: extension of a file path to support in the registered method

version::Union{AbstractString,Nothing}: version of the registered methods

Examples

julia> using DataFrames

julia> extension = "jdf";

julia> version = nothing;

julia> df = DataFrame(Dict("a" => 1:5, "b" => 6:10))
5×2 DataFrame
 Row │ a      b
     │ Int64  Int64
─────┼──────────────
   1 │     1      6
   2 │     2      7
   3 │     3      8
   4 │     4      9
   5 │     5     10

julia> filepath = joinpath(tempdir(), "file.$extension");

julia> register_default_jdf_file_handler(
            extension = extension;
            version = version,
        );

julia> save_file(filepath, df, version=version);

julia> loaded_df = load_file(filepath, version=version);

julia> loaded_df == df
true

DataWorkstation.IO.register_default_serialization_file_handler — Function

register_default_serialization_file_handler(
    ; extension = "slz", version = nothing) -> Nothing

Function to register methods for load_file() and save_file() with a default implementation based on the Serialization library, to be compatible with file paths having an extension value and some version. This internall calls register_file_handler().

Useful to save objects like machine learning models or just Julia objects to be retrieved in a Julia session, by using a native Julia implementation.

Arguments

extension::AbstractString: extension of a file path to support in the registered method

version::Union{AbstractString,Nothing}: version of the registered methods

Examples

julia> extension = "slz";

julia> version = nothing;

julia> obj = (;a=1, b=2);

julia> filepath = joinpath(tempdir(), "file.$extension");

julia> register_default_serialization_file_handler(
            extension = extension;
            version = version,
        );

julia> save_file(filepath, obj, version=version);

julia> loaded_obj = load_file(filepath, version=version);

julia> loaded_obj == obj
true

Logging utils

DataWorkstation.IO.custom_logger_meta_formatter — Function

custom_logger_meta_formatter(
    labels::Union{Vector{T}, Nothing} = nothing,
    log_level::Logging.LogLevel = Logging.Debug;
    sep::AbstractString = " | ",
    date_format::Union{AbstractString, Nothing} = "yyyy-mm-dd HH:MM:SS",
) where {T <: Union{Any, String}} -> Tuple{Symbol, String, String}

Function to allow a custom format for the logging messages to generate with a Logging.ConsoleLogger instance. The format is the following one: [ (date) (sep) (labels separated by 'sep') (log level): (message) For example: [ 2022-01-01 00:00:00 | tag=test | INFO: Happy new year!

Arguments

labels::Union{Vector{T}, Nothing}: labels to be used in the prefix of the logging messages. If nothing or an empty Array, no labels will be added to the message.

log_level::Logging.LogLevel: Level for the log message to print.

Keywords

sep::AbstractString: Separator for the labels to print in the message prefix

date_format::Union{AbstractString, Nothing}: Format for the timestamp to print in the message prefix. If nothing, no timestamp will be printed.

Returns

Tuple{Symbol, String, String}: A tuple of log level, prefix of message, and suffix of message to be used by a Logger instance when printing a logging message.

Examples

julia> using Logging

julia> custom_logger_meta_formatter(date_format=nothing)
(:blue, "DEBUG:", "")

julia> custom_logger_meta_formatter(["level", "sublevel"],
       Logging.Info, date_format=nothing, sep=" - ")
(:blue, "level - sublevel - INFO:", "")

DataWorkstation.IO.get_formatted_logger — Function

get_formatted_logger(
    logging_config::LoggingConfig,
    args...;
    stream::Base.IO=stderr,
    sep::String=" | ",
    date_format::AbstractString = "yyyy-mm-dd HH:MM:SS",
    kwargs...) -> Logging.ConsoleLogger

Get an instance of Logging.ConsoleLogger, having a format for the messages based on the given arguments.

This is a version of get_formatted_logger(), allowing to configure some parameters with logging_config.

Arguments

logging_config::LoggingConfig: Configuration of logging to implement, in particular the minimum level of log to print and the colors to use by logging level.

args...: Labels to be printed as in the message prefix.

Keywords

stream::Base.IO: Stream for the messages to print. Default to stderr.

sep::AbstractString: Separator for the labels to print in the message prefix.

date_format::Union{AbstractString, Nothing}: Format for the timestamp. to print in the message prefix. If nothing, no timestamp will be printed.

kwargs...: Labels to be printed as key=value in the message prefix.

Returns

Logging.ConsoleLogger: Logger formatted with custom_logger_meta_formatter().

Throws

ErrorException: In case the log level is not supported, or if some error occurs during the logger obtention.

julia> using Logging

julia> Logging.with_logger(get_formatted_logger(LoggingConfig(), scope="a", tag="b")) do
            @info "Hello world!"
        end
[ 2022-07-17 23:36:20 | scope=a | tag=b | INFO: Hello world!

julia> Logging.with_logger(
        get_formatted_logger(LoggingConfig(min_level=:Debug), "tag1", level="1")) do
            @debug "Logging in this level"
            Logging.with_logger(get_formatted_logger(:Debug, "tag2", level="2")) do
                @debug "Logging in another level"
            end
        end
[ 2022-07-17 23:37:38 | tag1 | level=1 | DEBUG: Logging in this level
[ 2022-07-17 23:37:38 | tag2 | level=2 | DEBUG: Logging in another level

get_formatted_logger(
    min_level::Symbol=:Debug,
    args...;
    stream::Base.IO=stderr,
    sep::String=" | ",
    date_format::AbstractString = "yyyy-mm-dd HH:MM:SS",
    colors_config::ColorsConfig = ColorsConfig(),
    kwargs...) -> Logging.ConsoleLogger

Get an instance of Logging.ConsoleLogger, having a format for the messages based on the given arguments.

This is useful for the package to log the operations on a workflow with details about in which level they occur, which is important during debugging.

Arguments

min_level::Symbol: Minimum level of log to print, represented as as symbol (e.g. :Debug, :Info, :Error, :Warn).

args...: Labels to be printed as in the message prefix.

Keywords

stream::Base.IO: Stream for the messages to print. Default to stderr.

sep::AbstractString: Separator for the labels to print in the message prefix.

date_format::Union{AbstractString, Nothing}: Format for the timestamp. to print in the message prefix. If nothing, no timestamp will be printed.

kwargs...: Labels to be printed as key=value in the message prefix.

Returns

Logging.ConsoleLogger: Logger formatted with custom_logger_meta_formatter().

Throws

ErrorException: In case the log level is not supported, or if some error occurs during the logger obtention.

julia> using Logging

julia> Logging.with_logger(get_formatted_logger(scope="repl", tag="test")) do
            @info "Hello world!"
        end
[ 2022-05-18 17:27:01 | scope=repl | tag=test | INFO: Hello world!

julia> Logging.with_logger(get_formatted_logger(:Debug, "tag1", level="1")) do
            @debug "Logging in this level"
            Logging.with_logger(get_formatted_logger(:Debug, "tag2", level="2")) do
                @debug "Logging in another level"
            end
        end
[ 2022-05-18 17:27:16 | tag1 | level=1 | DEBUG: Logging in this level
[ 2022-05-18 17:27:16 | tag2 | level=2 | DEBUG: Logging in another level

DataWorkstation.IO.ColorsConfig — Type

ColorsConfig(colors_by_level::Dict{Symbol,Symbol})
ColorsConfig()

Abstraction to handle the configuration of colors for logging operations. If no arguments specified, using a default configuration.

Arguments

colors_by_level::Dict{Symbol,Symbol}: Mapping of logging level to color (i.e. Dict(:Info => :blue, :Debug => :grey, :Error => :red))

Throws

ErrorException: In case some logging level defined is not supported.

Workflows module

Logging utils

DataWorkstation.Workflows.LoggingConfig — Type

LoggingConfig(;
    min_level::Symbol=:Info,
    colors_config::ColorsConfig=ColorsConfig(),
    workflow::Bool=true, jobs::Bool=true, steps::Bool=true,
)
LoggingConfig(workflow_spec::WorkflowSpec)

Abstraction to manage configuration relevant for the logging of a workflow to execute.

Fields

min_level::Symbol: Minimum level of log to print, represented as as symbol (e.g. :Debug, :Info, :Error, :Warn).

colors_config::ColorsConfig: Configuration of colors to use in console logging, through an instance of a ColorsConfig.

workflow::Bool: Flag to indicate that logging at workflow level should be printed.

jobs::Bool: Flag to indicate that logging at job level should be printed.

steps::Bool: Flag to indicate that logging at step level should be printed.

Arguments

workflow_spec::WorkflowSpec: Specification of the workflow, so the logging configuration can be extracted from this WorkflowSpec instance.

```

Step functions

DataWorkstation.Workflows.register_step_function — Function

register_step_function(step_sym::Symbol, func::Function; overwrite::Bool = true)
register_step_function(func::Function; overwrite::Bool = true)

Register a step function which is expected to be runned as part of some workflow specified through a WorkflowSpec.

A step function is just a simple Function that can only take a ConfigObject instance as unique argument to operate, which makes simpler to build workflows that chain step-by-step operations with some configured behavior. This kind of function can be identified through a step_sym symbol, or with a function that will be identified with its name. To be a valid function, it should meet the following aspects:

It must be available for calls done outside its origin module (so make sure it can be

imported, or just use eval when declaring the function).

It must have one argument of the type ConfigObject or Any. This object

should contain all the necessary configuration to allow the step work correctly.

Its return type will depend on the design of the involved step function,

but in general it is expected to be Nothing.

Arguments

step_sym::Symbol: Identifier for the step function to register.

func::Function: Function to register.

overwrite::Bool = true: In case the step function was already registered, decide if

it should be overwriten.

Throws

ErrorException: In case the step function is not valid, or it already exists

(and overwrite = false).

julia> function print_config(cfg) println("Config: $cfg") end
print_config (generic function with 1 method)

julia> register_step_function(:PrintConfig, print_config)
run_step_function (generic function with 2 methods)

julia> run_step_function(:PrintConfig, ConfigObject((;a=1)))
# prints: "Config: ConfigObject((a = 1,))"

julia> register_step_function(:GetConfiguredName, cfg -> cfg.name)
run_step_function (generic function with 3 methods)

julia> run_step_function(:GetConfiguredName, ConfigObject((;name="pepe")))
"pepe"

julia> multiply_x_and_y(cfg) = cfg.x * cfg.y
multiply_x_and_y (generic function with 1 method)

julia> register_step_function(multiply_x_and_y)
run_step_function (generic function with 4 methods)

julia> run_step_function(:multiply_x_and_y, ConfigObject((;x=5, y=4)))
20

DataWorkstation.Workflows.run_step_function — Function

run_step_function(s::Symbol, cfg::ConfigObject) -> Any

Run a step function previously registered through a call of register_step_function(). The return type will depend on the design of the involved step function, but in general it is expected to be Nothing.

This is an internal method which is convenient for the execution of a workflow through run_workflow(), so it should be used directly only for testing purposes.

For examples of usage, look at the documentation of register_step_function().

Arguments

s::Symbol: Identifier for the step function.

cfg::ConfigObject: Argument for the step function to run.

Throws

ErrorException: If some error occurs during the step function execution.

DataWorkstation.Workflows.is_valid_step_function — Function

is_valid_step_function(func::Function) -> Bool

Returns true if the function is considered a valid step function for a Workflow, so it can be registered through a call of register_step_function().

The function must have only one argument of the type ConfigObject or Any. This object should contain all the necessary configuration to allow the step work correctly.

Arguments

func::Function: Function to be validated.

Returns

Bool: Returns true if the function is considered a valid step function.

Examples

julia> is_valid_step_function((x1, x2) -> x1 + x2)
false

julia> is_valid_step_function(a::Int -> a + 1)
false

julia> is_valid_step_function(cfg::ConfigObject -> cfg.x)
true

julia> is_valid_step_function(cfg -> cfg.x)
true

Workflow specification and running

DataWorkstation.Workflows.WorkflowSpec — Type

WorkflowSpec(nt::NamedTuple)
WorkflowSpec(filename::String)

Abstraction to specify the properties of a Workflow to be executed. The specification can be done either through a NamedTuple or a TOML file. It can be used just like a NamedTuple to access to its properties.

Fields

_spec::NamedTuple: stores the keys and values for the specification.

Examples

julia> raw_nt = (;a=1, b=(;c=3, d=4))
(a = 1, b = (c = 3, d = 4))

julia> spec = WorkflowSpec(raw_nt)
WorkflowSpec((a = 1, b = (c = 3, d = 4)))

julia> spec.b
(c = 3, d = 4)

julia> spec.b.d
4

julia> length(spec)
2

julia> keys(spec), values(spec)
((:a, :b), (1, (c = 3, d = 4)))

julia> collect(spec)
(a = 1, b = (c = 3, d = 4))

julia> using TOML

julia> raw_dict = Dict("a" => 1, "b" => Dict("c" => 3, "d" => 4))
Dict{String, Any} with 2 entries:
  "b" => Dict("c"=>3, "d"=>4)
  "a" => 1

julia> filename = tempdir() * "/workflow.toml";

julia> open(filename, "w") do io
       TOML.print(io, raw_dict)
       end;

julia> spec == WorkflowSpec(filename)
true

DataWorkstation.Workflows.run_workflow — Function

run_workflow(cfg, workflow_spec; jobs_keys) -> Float
run_workflow(cfg, filename; jobs_keys) -> Float
run_workflow(filename; jobs_keys, config_root) -> Float

Run a selection of jobs from a workflow specified, each of them having step functions previously registered through calls of register_step_function(), using the configuration specified in a ConfigObject.

Arguments

cfg::ConfigObject: Configuration to use during steps running.

workflow_spec::WorkflowSpec: Specification for the given workflow.

filename::String: File to load workflow (and configuration) specification.

jobs_keys::Vector{Symbol}: Selection of jobs to run from the given workflow.

config_root::Union{AbstractString,Nothing}: Root for the configuration file to load (optional)

Returns

Float: duration of the execution.

Throws

ErrorException: If some WorkflowSpec entry needed does not exist, or some error occurs during the jobs execution.

Examples

julia> cfg = ConfigObject((;
           result_job_a_step_1 = "res_job_a_step_1",
           result_job_a_step_2 = "res_job_a_step_2",
           result_job_b_step_1 = "res_job_b_step_1",
           result_job_b_step_2 = "res_job_b_step_2",
       ))
ConfigObject((result_job_a_step_1 = "res_job_a_step_1",
    result_job_a_step_2 = "res_job_a_step_2", result_job_b_step_1 = "res_job_b_step_1",
    result_job_b_step_2 = "res_job_b_step_2"))

julia> workflow_spec = WorkflowSpec((;
        name = "test_workflow",
        jobs = (;
            job_a_test_run_jobs = (; steps = ["job_a_step_1", "job_a_step_2"]),
            job_b_test_run_jobs = (; steps = ["job_b_step_1", "job_b_step_2"],
                                    needs = "job_a_test_run_jobs"),
        ),
        ))
WorkflowSpec((name = "test_workflow", jobs = (job_a_test_run_jobs =
    (steps = ["job_a_step_1", "job_a_step_2"],), job_b_test_run_jobs =
    (steps = ["job_b_step_1", "job_b_step_2"], needs = "job_a_test_run_jobs"))))

julia> result = []  # Just a collector of results from runned steps
Any[]

julia> job_a_step_1(c::ConfigObject) = push!(result, c.result_job_a_step_1)
job_a_step_1 (generic function with 1 method)

julia> job_a_step_2(c::ConfigObject) = push!(result, c.result_job_a_step_2)
job_a_step_2 (generic function with 1 method)

julia> job_b_step_1(c::ConfigObject) = push!(result, c.result_job_b_step_1)
job_b_step_1 (generic function with 1 method)

julia> job_b_step_2(c::ConfigObject) = push!(result, c.result_job_b_step_2)
job_b_step_2 (generic function with 1 method)

julia> register_step_function(job_a_step_1)
run_step_function (generic function with 2 methods)

julia> register_step_function(job_a_step_2)
run_step_function (generic function with 3 methods)

julia> register_step_function(job_b_step_1)
run_step_function (generic function with 4 methods)

julia> register_step_function(job_b_step_2)
run_step_function (generic function with 5 methods)

julia> run_workflow(cfg, workflow_spec);

julia> result
4-element Vector{Any}:
 "res_job_a_step_1"
 "res_job_a_step_2"
 "res_job_b_step_1"
 "res_job_b_step_2"